2.6 refactor; move nn modules to separate files
move traffic sign classification (recognition) - and speed recommendation nn modules to separate files
